Create object dir
