+++ object/point_light.h
@@ -6,7 +6,7 @@
 #include "vec3.h"
 #include "ray.h"
 #include "material.h"
-#include "object.h"
+#include "object/object.h"
 
 namespace tsukihi {
