+++ object/raymarching_mbox.h
@@ -9,7 +9,7 @@
 #include "material.h"
 #include "constant.h"
 #include "intersection.h"
-#include "raymarching_object.h"
+#include "object/raymarching_object.h"
 
 namespace tsukihi {
 	struct RaymarchingMbox : public RaymarchingObject {
