+++ object/raymarching_menger_sponge.h
@@ -9,7 +9,7 @@
 #include "material.h"
 #include "constant.h"
 #include "intersection.h"
-#include "raymarching_object.h"
+#include "object/raymarching_object.h"
 
 namespace tsukihi {
