+++ object/raymarching_sphere.h
@@ -7,7 +7,7 @@
 #include "material.h"
 #include "constant.h"
 #include "intersection.h"
-#include "raymarching_object.h"
+#include "object/raymarching_object.h"
 
 namespace tsukihi {
