+++ tsukihi.vcxproj
@@ -128,20 +128,20 @@
     <ClInclude Include="intersection.h" />
     <ClInclude Include="material.h" />
     <ClInclude Include="math.h" />
-    <ClInclude Include="object.h" />
-    <ClInclude Include="point_light.h" />
+    <ClInclude Include="object\object.h" />
+    <ClInclude Include="object\point_light.h" />
     <ClInclude Include="image.h" />
     <ClInclude Include="pathtracing_renderer.h" />
     <ClInclude Include="random.h" />
     <ClInclude Include="ray.h" />
-    <ClInclude Include="raymarching_mbox.h" />
-    <ClInclude Include="raymarching_menger_sponge.h" />
-    <ClInclude Include="raymarching_menger_sponge_loop.h" />
-    <ClInclude Include="raymarching_object.h" />
-    <ClInclude Include="raymarching_sphere.h" />
+    <ClInclude Include="object\raymarching_mbox.h" />
+    <ClInclude Include="object\raymarching_menger_sponge.h" />
+    <ClInclude Include="object\raymarching_menger_sponge_loop.h" />
+    <ClInclude Include="object\raymarching_object.h" />
+    <ClInclude Include="object\raymarching_sphere.h" />
     <ClInclude Include="renderer.h" />
     <ClInclude Include="scene.h" />
-    <ClInclude Include="sphere.h" />
+    <ClInclude Include="object\sphere.h" />
     <ClInclude Include="lib\stb_image.h" />
     <ClInclude Include="lib\stb_image_write.h" />
     <ClInclude Include="vec3.h" />
